Reshape expressions longer or wider

addin_reshape() lets you cycle between different shapes of function calls. For instance, reshaping transforms code from wide to long shape and vice versa:

list(a, b, c)

list(
  a,
  b,
  c
)

Note that for function definitions, addin_reshape() cycles through two different long shapes. The traditional L form uses more horizontal space whereas the flat form uses less horizontal space and the arguments are always aligned at double indent:

foo <- function(a, b, c) {
  NULL
}

foo <- function(a,
                b,
                c) {
  NULL
}

foo <- function(
  a,
  b,
  c
) {
  NULL
}

addin_reshape()